Move OcNewsUrlReader into a separate file
